refactor stringy generate sql fn from sql into templates and snapshots
