+++ .eslintignore
@@ -1,1 +1,5 @@
 lib
+
+src/templates
+
+snapshots
