+++ tsconfig.json
@@ -8,5 +8,5 @@
     "declaration": true,
   },
   "include": ["./src/**/*"],
-  "exclude": ["./src/**/*.test.ts"]
+  "exclude": ["./src/**/*.test.ts", "./src/templates/*"]
 }
